pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-multi-engineer-agent"
version = "0.1.0"
description = "Multi-model engineering agent using AWS Strands across OpenAI, Anthropic, and Ollama"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "openai>=1.35.0",
    "anthropic>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
agent = "agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["agent", "providers", "tools", "tasks", "eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
